requirements.txt
altair
pandas
streamlit
plotly-express
mysql-connector-python
sqlalchemy
pymysql
cryptography
